requirements.txt
altair==5.5.0
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autobahn==19.11.2
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-admin-sortable2==2.2.8
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.9
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.0.3
fonttools==4.60.1
fqdn==1.5.1
freetype-py==2.5.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
html2docx==1.6.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiteconnect==5.0.1
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mongoengine==0.29.1
more-itertools==10.8.0
mysqlclient==2.2.7
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
num2words==0.5.14
numpy==2.2.4
openpyxl==3.1.5
oscrypto==1.3.0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.3.6
premailer==3.10.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.4
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pycryptodomex==3.23.0
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pymongo==4.14.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pypdf==6.1.1
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-json-logger==3.3.0
pytz==2025.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzipper==0.3.6
pyzmq==26.2.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlPyCairo==0.4.0
rpds-py==0.24.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.44.0
streamlit-option-menu==0.4.0
svglib==1.6.0
tablib==3.8.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritools==5.0.0
urllib3==2.3.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.13
weasyprint==66.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.13
xhtml2pdf==0.2.17
zope.interface==7.2
zopfli==0.2.3.post1
